+++ framework/modules/ecommerce/controllers/eventregistrationController.php
@@ -352,9 +352,12 @@ class eventregistrationController extends expController {
 
     public function delete_registrant() {
         global $db;
-        $connector_id = $this->params['connector_id'];
 
-        $db->delete("eventregistration_registrants", "connector_id='{$connector_id}'");
+        $registrant = $db->selectObject("eventregistration_registrants", "id='{$this->params['id']}'");
+        $event = $db->selectObject("eventregistration", "id='{$registrant->event_id}'");
+        $db->delete("eventregistration_registrants", "id='{$this->params['id']}'");
+        $event->number_of_registrants = $db->countObjects("eventregistration_registrants", "event_id='{$event->id}'");
+        $db->updateObject($event,"eventregistration");
         flash('message', gt("Registrant successfully deleted."));
         expHistory::back();
     }
@@ -362,67 +365,98 @@ class eventregistrationController extends expController {
     public function edit_registrant() {
         global $db;
 
-        $event_id = $this->params['event_id'];
-        $connector_id = @$this->params['connector_id'];
-        if (empty($connector_id)) {
-            $connector_id = "admin-created" . rand() . time(); //Meaning it is been added by admin
-        }
-        $reg_data = $db->selectObjects("eventregistration_registrants", "connector_id ='{$connector_id}'");
+//        $event_id     = $this->params['event_id'];
+//        $connector_id = @$this->params['connector_id'];
+//        if (empty($connector_id)) {
+//            $connector_id = "admin-created" . rand() . time(); //Meaning it is been added by admin
+//        }
+//        $reg_data   = $db->selectObjects("eventregistration_registrants", "connector_id ='{$connector_id}'");
+
         $registrant = array();
-        foreach ($reg_data as $item) {
-            $registrant[$item->control_name] = $item->value;
+        if (!empty($this->params['id'])) {
+            $reg_data   = $db->selectObject("eventregistration_registrants", "id ='{$this->params['id']}'");
+  //        foreach ($reg_data as $item) {
+  //            $registrant[$item->control_name] = $item->value;
+  //        }
+            $registrant = expUnserialize($reg_data->value);
+            $registrant['id'] = $reg_data->id;
+            $eventid = $reg_data->event_id;
+        } else {
+            $eventid = $this->params['event_id'];
         }
 
-        $event = new eventregistration($event_id);
+        $event = new eventregistration($eventid);
 
         // eDebug($registrant, true);
         assign_to_template(array(
-            'registrant'   => $registrant,
-            'event'        => $event,
-            'connector_id' => $connector_id
+            'registrant'=> $registrant,
+            'event'=> $event,
+//            'connector_id' => $connector_id
         ));
     }
 
     public function update_registrant() {
         global $db;
-        $event_id = $this->params['event_id'];
-        $connector_id = $this->params['connector_id'];
-        $fields = $this->params['event'];
-        $obj = '';
-        // eDebug($fields);
-        foreach ($fields as $key => $value) {
-            $obj = $db->selectObject("eventregistration_registrants", "event_id = '{$event_id}' AND connector_id = '{$connector_id}' AND control_name='{$key}'");
 
-            if (!empty($obj)) {
-                $obj->value = $value;
-                if (!empty($value)) {
-                    $db->updateObject($obj, 'eventregistration_registrants');
-                } else {
-                    $db->delete('eventregistration_registrants', 'id=' . $obj->id);
-                }
-            } else {
-                if (!empty($value)) {
-                    $reg = new stdClass();
-                    $reg->event_id = $event_id;
-                    $reg->control_name = $key;
-                    $reg->value = $value;
-                    $reg->connector_id = $connector_id;
-                    $reg->registered_date = time();
+//        $event_id     = $this->params['event_id'];
+//        $connector_id = $this->params['connector_id'];
+//        $fields       = $this->params['event'];
+//        $obj          = '';
+//        // eDebug($fields);
+//        foreach ($fields as $key => $value) {
+//            $obj = $db->selectObject("eventregistration_registrants", "event_id = '{$event_id}' AND connector_id = '{$connector_id}' AND control_name='{$key}'");
+//
+//            if (!empty($obj)) {
+//                $obj->value = $value;
+//                if (!empty($value)) {
+//                    $db->updateObject($obj, 'eventregistration_registrants');
+//                } else {
+//                    $db->delete('eventregistration_registrants', 'id=' . $obj->id);
+//                }
+//            } else {
+//                if (!empty($value)) {
+//                    $reg                  = '';
+//                    $reg->event_id        = $event_id;
+//                    $reg->control_name    = $key;
+//                    $reg->value           = $value;
+//                    $reg->connector_id    = $connector_id;
+//                    $reg->registered_date = time();
+//
+//                    $db->insertObject($reg, 'eventregistration_registrants');
+//                }
+//
+//            }
+//        }
 
-                    $db->insertObject($reg, 'eventregistration_registrants');
-                }
-
-            }
+        $registrant = $db->selectObject("eventregistration_registrants", "id ='{$this->params['id']}'");
+        $data = array();
+        $data['name'] = $this->params['name'];
+        $data['email'] = $this->params['email'];
+        $data['phone'] = $this->params['phone'];
+        if (empty($registrant->id)) {
+            $registrant = new stdClass();
+            $registrant->event_id = $this->params['event_id'];
+            $registrant->control_name = 0;
+            $registrant->value = serialize($data);
+            $registrant->connector_id = 'admin-created';
+            $registrant->registered_date = time();
+            $db->insertObject($registrant,"eventregistration_registrants");
+            $event = $db->selectObject("eventregistration", "id='{$registrant->event_id}'");
+            $event->number_of_registrants = $db->countObjects("eventregistration_registrants", "event_id='{$event->id}'");
+            $db->updateObject($event,"eventregistration");
+        } else {
+            $registrant->value = serialize($data);
+            $db->updateObject($registrant,"eventregistration_registrants");
         }
-        // exit();
-        redirect_to(array('controller' => 'eventregistration', 'action' => 'view_registrants', 'id' => $event_id));
+
+        redirect_to(array('controller'=> 'eventregistration', 'action'=> 'view_registrants', 'id'=> $registrant->event_id));
     }
 
     public function export() {
         global $db;
 
-        $event = new eventregistration($this->params['id']);
-        $sql = "SELECT connector_id FROM " . DB_TABLE_PREFIX . "_eventregistration_registrants GROUP BY connector_id";
+        $event              = new eventregistration($this->params['id']);
+        $sql                = "SELECT connector_id FROM " . DB_TABLE_PREFIX . "_eventregistration_registrants GROUP BY connector_id";
         $order_ids_complete = $db->selectColumn("eventregistration_registrants", "connector_id", "connector_id <> '0' AND event_id = {$event->id}", "registered_date", true);
 
         foreach ($order_ids_complete as $item) {
@@ -432,9 +466,9 @@ class eventregistrationController extends expController {
             }
         }
 
-        $header = array();
+        $header        = array();
         $control_names = array();
-        $header[] = '"Date Registered"';
+        $header[]      = '"Date Registered"';
         //Check if it has ticket types
         if ($event->hasOptions()) {
             $header[] = '"Ticket Types"'; //Add some configuration here
@@ -466,14 +500,15 @@ class eventregistrationController extends expController {
         }
 
         // new method to check for guests/registrants
-        if (!empty($event->num_guest_allowed)) {
+//        if (!empty($event->num_guest_allowed)) {
+        if (!empty($event->quantity)) {
             $registered = array();
             if (!empty($order_ids)) foreach ($order_ids as $order_id) {
                 $newregistrants = $db->selectObjects("eventregistration_registrants", "connector_id ='{$order_id}'");
-                $registered = array_merge($registered, $newregistrants);
+                $registered = array_merge($registered,$newregistrants);
             }
 //            $registrants = array();
-            foreach ($registered as $key => $person) {
+            foreach ($registered as $key=>$person) {
                 $registered[$key]->person = expUnserialize($person->value);
             }
             $header[] = '"Name"';
@@ -482,12 +517,12 @@ class eventregistrationController extends expController {
         }
 
         if (LANG_CHARSET == 'UTF-8') {
-            $out = chr(0xEF) . chr(0xBB) . chr(0xBF); // add utf-8 signature to file to open appropriately in Excel, etc...
+            $out = chr(0xEF).chr(0xBB).chr(0xBF);  // add utf-8 signature to file to open appropriately in Excel, etc...
         } else {
             $out = "";
         }
-        $out .= implode(",", $header);
-        $out .= "\n";
+        $out  .= implode(",", $header);
+        $out  .= "\n";
         $body = '';
 //        foreach ($order_ids as $order_id) {
 //            $body .= '"' . date("M d, Y h:i a", $db->selectValue("eventregistration_registrants", "registered_date", "event_id = {$event->id} AND connector_id = '{$order_id}'")) . '",';
@@ -526,14 +561,14 @@ class eventregistrationController extends expController {
 //        fputs($f, $out);
 //        fclose($f);
 
-        // CREATE A TEMP FILE
-        $tmpfname = tempnam(getcwd(), "rep"); // Rig
+		// CREATE A TEMP FILE
+		$tmpfname = tempnam(getcwd(), "rep"); // Rig
 
-        $handle = fopen($tmpfname, "w");
-        fwrite($handle, $out);
-        fclose($handle);
+		$handle = fopen($tmpfname, "w");
+		fwrite($handle,$out);
+		fclose($handle);
 
-        if (file_exists($tmpfname)) {
+		if(file_exists($tmpfname)) {
             // NO buffering from here on out or things break unexpectedly. - RAM
             ob_end_clean();
 
@@ -542,9 +577,9 @@ class eventregistrationController extends expController {
             // MSIE and Opera seems to prefer 'application/octetstream'
             // It seems that other headers I've added make IE prefer octet-stream again. - RAM
             $mime_type = (EXPONENT_USER_BROWSER == 'IE' || EXPONENT_USER_BROWSER == 'OPERA') ? 'application/octet-stream;' : 'text/comma-separated-values;';
-            header('Content-Type: ' . $mime_type . ' charset=' . LANG_CHARSET . "'");
+            header('Content-Type: ' . $mime_type . ' charset=' . LANG_CHARSET. "'");
             header('Expires: ' . gmdate('D, d M Y H:i:s') . ' GMT');
-            header("Content-length: " . filesize($tmpfname));
+            header("Content-length: ".filesize($tmpfname));
             header('Content-Transfer-Encoding: binary');
             header('Content-Encoding:');
             header('Content-Disposition: attachment; filename="' . $fn . '";');
@@ -559,13 +594,13 @@ class eventregistrationController extends expController {
 
             readfile($tmpfname);
             if (DEVELOPMENT == 0)
-                exit();
+            exit();
         }
     }
 
     public function get_guest_controls($ajax = '') {
-        $id = $this->params['id'];
-        $ctr = $this->params['counter'];
+        $id    = $this->params['id'];
+        $ctr   = $this->params['counter'];
         $event = new eventregistration($id);
 
         $str = "";
@@ -595,9 +630,9 @@ class eventregistrationController extends expController {
             }
         }
 
-        $header = array();
+        $header        = array();
         $control_names = array();
-        $header[] = 'Date Registered';
+        $header[]      = 'Date Registered';
         //Check if it has ticket types
         if ($event->hasOptions()) {
             $header[] = "Types"; //Add some configuration here
@@ -629,29 +664,30 @@ class eventregistrationController extends expController {
         }
 
         // new method to check for guests/registrants
-        if (!empty($event->num_guest_allowed)) {
+//        if (!empty($event->num_guest_allowed)) {
+        $registrants = array();
+        if (!empty($event->quantity)) {
             $registered = array();
             if (!empty($order_ids)) foreach ($order_ids as $order_id) {
                 $newregistrants = $db->selectObjects("eventregistration_registrants", "connector_id ='{$order_id}'");
-                $registered = array_merge($registered, $newregistrants);
+                $registered = array_merge($registered,$newregistrants);
             }
-            $registrants = array();
             foreach ($registered as $person) {
-                $registrants[] = expUnserialize($person->value);
+                $registrants[$person->id] = expUnserialize($person->value);
             }
         }
 
         //Get the data and registrant emails
-        $email = array();
+        $email               = array();
         $num_of_guest_fields = 0;
-        $num_of_guest = 0;
-        $num_of_guest_total = 0;
+        $num_of_guest        = 0;
+        $num_of_guest_total  = 0;
 
         $body = array();
         if (!empty($order_ids)) foreach ($order_ids as $order_id) {
             $body[$order_id][] = date("M d, Y h:i a", $db->selectValue("eventregistration_registrants", "registered_date", "event_id = {$event->id} AND connector_id = '{$order_id}'"));
             if ($event->hasOptions()) {
-                $or = new order($order_id);
+                $or        = new order($order_id);
                 $orderitem = new orderitem();
                 if (isset($or->orderitem[0])) {
                     $body[$order_id][] = $orderitem->getOption($or->orderitem[0]->options);
@@ -660,7 +696,7 @@ class eventregistrationController extends expController {
                 }
             }
             foreach ($control_names as $control_name) {
-                $value = $db->selectValue("eventregistration_registrants", "value", "event_id = {$event->id} AND control_name ='{$control_name}' AND connector_id = '{$order_id}'");
+                $value             = $db->selectValue("eventregistration_registrants", "value", "event_id = {$event->id} AND control_name ='{$control_name}' AND connector_id = '{$order_id}'");
                 $body[$order_id][] = $value;
                 if (expValidator::isValidEmail($value) === true) {
                     $email[$value] = $value;
@@ -692,8 +728,8 @@ class eventregistrationController extends expController {
 
         $event->number_of_registrants = $registered;
         assign_to_template(array(
-            'event'       => $event,
-            'registrants' => $registrants,
+            'event'=> $event,
+            'registrants'=> $registrants,
 //            'header'=> $header,
 //            'body'=> $body,
 //            'email'=> $email
@@ -708,34 +744,53 @@ class eventregistrationController extends expController {
         }
 
         if (empty($this->params['email_subject'])) {
-            flash('error', gt('Please enter your email subject.'));
+            flash('error', gt('Nothing to Send! Please enter subject and message.'));
             expHistory::back();
         }
 
-        $email_arr = explode("|!|", $this->params['email_addresses']);
-        $email_subject = $this->params['email_subject'];
-        $email_message = $this->params['email_message'];
+//        $email_arr     = explode("|!|", $this->params['email_addresses']);
+        $email_addy = array_flip(array_flip($this->params['email_addresses']));
+        $email_addy = array_map('trim', $email_addy);
+        $email_addy = array_filter($email_addy);
+
+        $headers = array(
+            "MIME-Version" => "1.0",
+            "Content-type" => "text/html; charset=" . LANG_CHARSET
+        );
 
         $mail = new expMail();
 
-        foreach ($this->params['expFile']['attachments'] as $attach) {
-            $expFile = new expFile($attach);
-            if (!empty($expFile->id)) {
-                $mail->attach_file_on_disk($expFile->path, $expFile->mimetype);
-            }
+//        foreach ($this->params['expFile']['attachments'] as $attach) {
+//            $expFile = new expFile($attach);
+//            if (!empty($expFile->id)) {
+//                $mail->attach_file_on_disk($expFile->path, $expFile->mimetype);
+//            }
+//        }
+        if (!empty($_FILES['attach']['size'])) {
+            $dir = 'tmp';
+            $filename = expFile::fixName(time().'_'.$_FILES['attach']['name']);
+            $dest = $dir.'/'.$filename;
+            //Check to see if the directory exists.  If not, create the directory structure.
+            if (!file_exists(BASE.$dir)) expFile::makeDirectory($dir);
+            // Move the temporary uploaded file into the destination directory, and change the name.
+            expFile::moveUploadedFile($_FILES['attach']['tmp_name'],BASE.$dest);
+            $finfo = finfo_open(FILEINFO_MIME_TYPE);
+//                $relpath = str_replace(PATH_RELATIVE, '', BASE);
+            $ftype = finfo_file($finfo, BASE.$dest);
+            finfo_close($finfo);
+            $mail->attach_file_on_disk(BASE.$dest,$ftype);
         }
 
-        foreach ($email_arr as $email_addy) {
-            $mail->quickSend(array(
-                'html_message' => $email_message,
-                'text_message' => str_replace("<br>", "\r\n", $email_message),
-                'to'           => $email_addy,
-                'from'         => ecomconfig::getConfig('from_address'),
-                'subject'      => $email_subject
-            ));
-        }
+        $mail->quickBatchSend(array(
+            	'headers'=>$headers,
+                'html_message'=> $this->params['email_message'],
+                'text_message'=> strip_tags(str_replace("<br>", "\r\n", $this->params['email_message'])),
+                'to'          => $email_addy,
+                'from'        => ecomconfig::getConfig('from_address'),
+                'subject'     => $this->params['email_subject']
+        ));
 
-        flash('message', gt("You're email has been successfully sent."));
+        flash('message', gt("You're email to event registrants has been sent."));
         expHistory::back();
     }