+++ framework/modules/ecommerce/controllers/storeController.php
@@ -104,7 +104,8 @@ class storeController extends expController {
 
     function __construct($src = null, $params = array()) {
         global $db, $router, $section, $user;
-        parent::__construct($src = null, $params);
+//        parent::__construct($src = null, $params);
+        parent::__construct($src, $params);
 
         // we're setting the config here globally
         $this->grabConfig();